Move AttributeValue to common
